move class promenade.jvm.StacklessExceptionInfo to promenade.util package
